script for replies twilio->patient triggered by sms patient->twilio
This script has to be run in a ngrok tunnel or similar. It waits if a patient/client will send an SMS to the twilio number of the service (patient->twilio). When this happens the script checks the content and if it finds keywords matching the logic of the script (e.g. "MEDICATIONS") an SMS is being sent back (twilio->patient) with a message corresponding to the logic (e.g. "Your logged medications are: aaaaa, bbbbb, .....")